DEMOS-613: moved demonstrationDetail hook to within component.
